v2.2.64+109
Payment v2 tested successfully
Added new payment model

Updated to rewards v2
Updated to transaction history v2
updated number utils

+++ lib/views/wallet/wallet_features/add_gas/add_gas_viewmodel.dart
@@ -208,8 +208,9 @@ class AddGasViewModel extends FutureViewModel {
     utxosNeeded = calculateUtxosNeeded(totalAmount, utxos);
     var fee = (utxosNeeded) * feePerInput + (2 * 34 + 10) * satoshisPerBytes;
     transFee = ((Decimal.parse(extraAmount.toString()) +
-            Decimal.parse(fee.toString()) / Decimal.parse('1e8')))
-        .toDouble();
+            (Decimal.parse(fee.toString()) / Decimal.parse('1e8')).toDecimal())
+        .toDouble());
+
     totalAmount = totalAmount + transFee;
     utxosNeeded = calculateUtxosNeeded(totalAmount, utxos);
     bool isRequiredUtxoValueIsMore = totalUtxos < utxosNeeded;
